improved hand detection
Hand detection no longer requires a sorted hand.
this is a prerequisite for implementing sort-by-rank
and drag-and-drop order.  and drag-and-drop order is
a prerequisite for the Death card and many jokers
that operate on "first King in hand" etc.

The scoring method now starts by making a histogram
of card ranks, and hands are detected with two
functions: contains_straight() which detects runs,
and count() which detects matches.  It also detects
A,2,3,4,5 straights, and is done in a way that will
make the Four Fingers joker easy to implement.

Also added five-of-a-kind, flush five, and flush house,
but not the planet cards for them.

Still detects Royal Flush as a distinct hand, but in
a simpler way.