optimized planet cards
the case changes made the diff look bad, so I thought
why not reduce some repetition in the planet cards by
giving them all the same function and description.

This required me to add support of description as
functions in addition to strings, which wasn't so hard.
This will allow future jokers to change over time and
mark the changes in the description.