can use mouse to rearrange hand
I found a way to reuse the insertion sort on arbitrary properties.
But I need it to sort ascending instead of descending, so I changed
the order property from 14-i to i in the loop that sets it, so the
hand still initialized in descending rank order.

I didn't have to program any extra animation because the place()
method already does that. The trick is to set pos_x and pos_y to the
mouse position directly, and then use place() to set it in motion
to where it belongs.

shouldn't be too hard to do the same thing with jokers, someday