animated descriptions for random jokers
when I refactored planets, I added the capability to
define description as a function, so it could be
parameterized.  But since the describe function is
called every frame the description is visible, I
realized I could use it to animate the descriptions.

also renamed them to "misprint" and "misprint chips"